requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
twilio==7.12.0
urllib3==1.26.11
